# packages/api-server/src/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: ConfigBuddy CMDB API
  version: 2.0.0
  description: |
    Comprehensive REST API for ConfigBuddy - Enterprise Configuration Management Database (CMDB) platform.

    ConfigBuddy v2.0 features:
    - Graph-based CI relationships with Neo4j
    - 37 connectors (17 TypeScript + 20 JSON-only)
    - Unified credential system with affinity matching
    - Discovery agent architecture with smart routing
    - AI-powered anomaly detection and impact analysis
    - Identity resolution and reconciliation

    ## Authentication

    All API endpoints (except `/api/v1/auth/login`) require authentication via:
    - **JWT Token**: Include `Authorization: Bearer <token>` header
    - **API Key**: Include `X-API-Key: <key>` header

    ## Rate Limiting

    Rate limits apply per authentication token/API key:
    - **Standard**: 1000 requests/hour
    - **Burst**: 100 requests/minute

    Rate limit headers included in responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  contact:
    name: ConfigBuddy Support
    url: http://localhost:8080
    email: support@configbuddy.io
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: http://staging-api.configbuddy.io/api/v1
    description: Staging server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /connectors/installed:
    get:
      tags: [Connectors]
      summary: List installed connectors
      description: Get all locally installed connectors
      operationId: getInstalledConnectors
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [discovery, connector]
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [name, connector_type, installed_at, updated_at]
            default: name
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Installed connectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstalledConnector'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /connectors/installed/{type}:
    get:
      tags: [Connectors]
      summary: Get installed connector
      description: Get details of an installed connector
      operationId: getInstalledConnectorDetails
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Connector details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstalledConnector'
        '404':
          description: Connector not installed
        '401':
          $ref: '#/components/responses/Unauthorized'
  /connectors/install:
    post:
      tags: [Connectors]
      summary: Install connector
      description: Install a connector from the registry
      operationId: installConnector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connector_type]
              properties:
                connector_type:
                  type: string
                  example: aws
                version:
                  type: string
                  pattern: '^\d+\.\d+\.\d+$'
                  example: 1.0.0
                force:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Connector installed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstalledConnector'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Connector already installed
  /connectors/{type}/update:
    put:
      tags: [Connectors]
      summary: Update connector
      description: Update connector to specific version
      operationId: updateConnector
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version:
                  type: string
                  pattern: '^\d+\.\d+\.\d+$'
                force:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Connector updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstalledConnector'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Connector not installed
  /connectors/{type}:
    delete:
      tags: [Connectors]
      summary: Uninstall connector
      description: Remove a connector from the system
      operationId: uninstallConnector
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Connector uninstalled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Connector not installed
  /connectors/{type}/verify:
    post:
      tags: [Connectors]
      summary: Verify connector
      description: Verify connector installation and integrity
      operationId: verifyConnector
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  errors:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Connector not installed
  /connectors/outdated:
    get:
      tags: [Connectors]
      summary: Check for updates
      description: Get list of outdated connectors
      operationId: checkOutdatedConnectors
      responses:
        '200':
          description: Outdated connectors
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    connector_type:
                      type: string
                    current_version:
                      type: string
                    latest_version:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /connectors/cache/refresh:
    post:
      tags: [Connectors]
      summary: Refresh registry cache
      description: Clear and refresh connector registry cache
      operationId: refreshRegistryCache
      responses:
        '200':
          description: Cache refreshed
        '401':
          $ref: '#/components/responses/Unauthorized'

  # =============================================================================
  # RECONCILIATION
  # =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key from /auth/api-keys endpoint
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation Error
            message: Invalid CI type provided
            details:
              - field: _type
                error: must be one of [server, virtual-machine, ...]
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or expired JWT token
...
  schemas:
...
    InstalledConnector:
      type: object
      properties:
        connector_type:
          type: string
        name:
          type: string
        version:
          type: string
        category:
          type: string
        enabled:
          type: boolean
        installed_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # =========================================================================
    # RECONCILIATION
    # =========================================================================
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error type/code
          example: Validation Error
        message:
          type: string
          description: Human-readable error message
          example: Invalid CI type provided
        details:
          type: array
          description: Detailed validation errors
          items:
            type: object
            properties:
              field:
                type: string
              error:
                type: string